Still trying to fix scorecard

+++ src/main_menu.c
@@ -129,7 +129,7 @@ static void main_menu_select_click(struct MenuLayer* menu, MenuIndex* cell_index
 
 void confirm_clear_scores_callback(bool result) {
   if (result) {
-    for (int i = 0; i < 18; i++) {
+    for (uint8_t i = 0; i < 18; i++) {
         set_my_strokes(i, 0);
       }
     clear_round_in_progress();
